+++ test/functionality/graphics/exp/circ.cpp
@@ -2,10 +2,10 @@
 #include "shape.h"
 
 namespace __zz_cib_ { namespace Graphics { namespace Circle {
-  static float __zz_cib_decl Area_0(const ::Graphics::Circle* __zz_cib_obj) {
+  static double __zz_cib_decl Area_0(const ::Graphics::Circle* __zz_cib_obj) {
     return __zz_cib_obj->Area();
   }
-  static float __zz_cib_decl Perimeter_1(const ::Graphics::Circle* __zz_cib_obj) {
+  static double __zz_cib_decl Perimeter_1(const ::Graphics::Circle* __zz_cib_obj) {
     return __zz_cib_obj->Perimeter();
   }
   static void __zz_cib_decl Draw_2(const ::Graphics::Circle* __zz_cib_obj, __zz_cib_HANDLE* ctx) {
