+++ test/functionality/graphics/exp/composite.h
@@ -23,11 +23,11 @@ typedef std::vector<Graphics::Shape*> Shapes;
     /**
      * @return Area of this shape object.
      */
-    virtual float Area() const;
+    virtual double Area() const;
     /**
      * @return Perimeter of this shape object.
      */
-    virtual float Perimeter() const;
+    virtual double Perimeter() const;
     /**
      * Draws this shape on a given device context.
      */
